Move threadpool xml parsing into own file
